requirements.txt
# requirements.txt
groq
google-generativeai
python-dotenv
pyaudio
pyttsx3
keyboard
pillow
piper-tts
faster-whisper
psutil
pyperclip
